+++ RELEASES.md
@@ -8,10 +8,21 @@ as-is, without any warranty. -->
 Version 1.3.0 (unreleased)
 ==========================
 
+  * The following methods were added to all fixed-point types and to
+    the `Fixed` trait:
+      * [`mul_add`][f-ma-1-3], [`checked_mul_add`][f-cma-1-3],
+        [`saturating_mul_add`][f-sma-1-3],
+        [`wrapping_mul_add`][f-wma-1-3],
+        [`overflowing_mul_add`][f-oma-1-3]
   * The new experimental feature [`unwrapped`][feat-un-1-3] was added,
     providing arithmetic methods that panic on overflow even when
     debug assertions are disabled.
 
+[f-cma-1-3]: https://tspiteri.gitlab.io/fixed/dev/fixed/struct.FixedI32.html#method.checked_mul_add
+[f-ma-1-3]: https://tspiteri.gitlab.io/fixed/dev/fixed/struct.FixedI32.html#method.mul_add
+[f-oma-1-3]: https://tspiteri.gitlab.io/fixed/dev/fixed/struct.FixedI32.html#method.overflowing_mul_add
+[f-sma-1-3]: https://tspiteri.gitlab.io/fixed/dev/fixed/struct.FixedI32.html#method.saturating_mul_add
+[f-wma-1-3]: https://tspiteri.gitlab.io/fixed/dev/fixed/struct.FixedI32.html#method.wrapping_mul_add
 [feat-un-1-3]: https://tspiteri.gitlab.io/fixed/dev/fixed/#experimental-optional-features
 
 Version 1.2.0 (2020-09-02)